addPoolsArray addVaultsArray added to initializer,
